src/gtk/transfer.c (check_done_process) - make sure the local file is not
2005-3-31 Brian Masney <masneyb@gftp.org>

        * src/gtk/transfer.c (check_done_process) - make sure the local file is
        not removed if the editor is set to a non-existent editor

+++ src/gtk/transfer.c
@@ -349,7 +349,8 @@ check_done_process (void)
                     {
                       ftp_log (gftp_logging_error, NULL,
                                _("Error: Child %d returned %d\n"), pid, procret);
-                      remove_file (ve_proc->filename);
+                      if (ve_proc->view)
+                        remove_file (ve_proc->filename);
                       continue;
                     }
                   else
